requirements.txt
langchain==0.3.0
langchain_community==0.3.0
langchain_core==0.3.5
langchain_groq==0.2.0
python-dotenv==1.0.1
streamlit==1.35.0
sentence-transformers==3.2.1
faiss-cpu==1.9.0